changed to identifying course by id not code
